# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Hacksim API
  version: 0.0.1
  description: >
    Backend for the Hacksim malware-detonation sandbox: triggers detonations, ingests guest
    telemetry, returns ATT&CK-mapped behavioural reports, and stores endpoint posture self-assessments.
servers:
  - url: http://localhost:5001
    description: Local development
paths:
  /api/start-simulation/{simulationType}:
    post:
      summary: Trigger a detonation (async)
      description: >
        Accepts the request, records a run, and detonates on a background worker. Requires an API key.
      security:
        - ApiKeyAuth: []
      parameters:
        - name: simulationType
          in: path
          required: true
          schema: { type: string, enum: [ransomware, worm, spyware, trojan] }
      responses:
        '202':
          description: Detonation accepted; runs in background
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DetonationAccepted' }
        '401':
          description: Missing or invalid API key
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiResponse' }
        '404':
          description: Unknown simulation type
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiResponse' }
...
  /api/runs/{runId}:
    get:
      summary: Single run status (for polling)
      parameters:
        - { name: runId, in: path, required: true, schema: { type: integer, format: int64 } }
      responses:
        '200':
          description: The run
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DetonationRun' }
        '404': { description: Run not found }
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
  schemas:
    ApiResponse:
      type: object
      properties:
        status: { type: string, enum: [success, error] }
        message: { type: string }
    DetonationAccepted:
      type: object
      properties:
        status: { type: string, example: success }
        message: { type: string }
        runId: { type: integer, format: int64 }
    DetonationRun:
      type: object
      properties:
        id: { type: integer, format: int64 }
        scenarioId: { type: string }
        status: { type: string, enum: [REQUESTED, TRIGGERED, FAILED] }
        clientIp: { type: string }
        message: { type: string, nullable: true }
        requestedAt: { type: string, format: date-time }
        completedAt: { type: string, format: date-time, nullable: true }
...